pyproject.toml
[project]
classifiers = [
  "Natural Language :: English",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
name = "mini-paint-catalog"
description = "A catalog of miniature paintings from different vendors in a structured format."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Jose M. Pizarro", email = "jose.pizarro-blanco@bam.de" },
]
maintainers = [
  { name = "Jose M. Pizarro", email = "jose.pizarro-blanco@bam.de" },
]
license = { file = "LICENSE" }
dependencies = [
  "click",
  "pydantic~=2.10.5",
  "requests",
  "playwright",
  "pillow",
  "cairosvg",
]

[project.urls]
"Homepage" = "https://github.com/JosePizarro3/mini-paint-catalog"
"Bug Tracker" = "https://github.com/JosePizarro3/mini-paint-catalog/issues"

[project.optional-dependencies]
dev = [
  "mypy==1.0.1",
  "ruff",
  "pytest",
  "pytest-timeout",
  "pytest-cov",
  "python-decouple",
  "structlog==24.4.0",
]
docu = [
  "mkdocs-material",
  "mkdocs-click",
  "mkdocstrings-python",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
include = ["*.py", "tests/*.py"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "dependencies",
]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "PL", # pylint
    "F", # Pyflakes
    "UP", # pyupgrade
    "I", # isort
]
ignore = [
    "F401", # Module imported but unused
    "E501", # Line too long ({width} > {limit} characters)
    "E701", # Multiple statements on one line (colon)
    "E731", # Do not assign a lambda expression, use a def
    "E402",  # Module level import not at top of file
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used instead of constant
    "PLW0603", # Using the global statement
    "PLW2901", # redefined-loop-name
    "PLR1714", # consider-using-in
    "PLR5501", # else-if-used
]
fixable = ["ALL"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use double quotes for strings.
quote-style = "double"
# indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
strict = false
ignore_missing_imports = true
follow_imports = "silent"
no_strict_optional = true
disable_error_code = "import, annotation-unchecked"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/minipaintcatalog/_version.py"

[tool.setuptools]
package-dir = { "" = "src" }
